requirements.txt
torch>=2.0.0
numpy>=1.21.0
faiss-cpu>=1.7.0
librosa>=0.10.0
soundfile>=0.12.0
spleeter>=2.3.0
flask>=2.3.0
pytest>=7.0.0
requests>=2.28.0
tqdm>=4.64.0
pydub>=0.25.0
matplotlib>=3.5.0
seaborn>=0.11.0
click>=8.0.0
